requirements.txt
aiohttp==3.9.5 ; python_version >= "3.11" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.11" and python_version < "4.0"
anyio==4.4.0 ; python_version >= "3.11" and python_version < "4.0"
async-timeout==4.0.3 ; python_version >= "3.11" and python_version < "3.12.0"
asyncpg==0.29.0 ; python_version >= "3.11" and python_version < "4.0"
attrs==23.2.0 ; python_version >= "3.11" and python_version < "4.0"
bcrypt==4.1.3 ; python_version >= "3.11" and python_version < "4.0"
black==24.4.2 ; python_version >= "3.11" and python_version < "4.0"
boto3==1.34.141 ; python_version >= "3.11" and python_version < "4.0"
botocore==1.34.141 ; python_version >= "3.11" and python_version < "4.0"
certifi==2024.6.2 ; python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.11" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and (sys_platform == "win32" or platform_system == "Windows")
dataclasses-json==0.6.7 ; python_version >= "3.11" and python_version < "4.0"
distro==1.9.0 ; python_version >= "3.11" and python_version < "4.0"
dnspython==2.6.1 ; python_version >= "3.11" and python_version < "4.0"
ecdsa==0.19.0 ; python_version >= "3.11" and python_version < "4.0"
email-validator==2.2.0 ; python_version >= "3.11" and python_version < "4.0"
fastapi-cli==0.0.4 ; python_version >= "3.11" and python_version < "4.0"
fastapi==0.111.0 ; python_version >= "3.11" and python_version < "4.0"
frozenlist==1.4.1 ; python_version >= "3.11" and python_version < "4.0"
greenlet==3.0.3 ; python_version < "3.13" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version >= "3.11"
h11==0.14.0 ; python_version >= "3.11" and python_version < "4.0"
httpcore==1.0.5 ; python_version >= "3.11" and python_version < "4.0"
httptools==0.6.1 ; python_version >= "3.11" and python_version < "4.0"
httpx==0.27.0 ; python_version >= "3.11" and python_version < "4.0"
idna==3.7 ; python_version >= "3.11" and python_version < "4.0"
iniconfig==2.0.0 ; python_version >= "3.11" and python_version < "4.0"
jinja2==3.1.4 ; python_version >= "3.11" and python_version < "4.0"
jmespath==1.0.1 ; python_version >= "3.11" and python_version < "4.0"
jsonpatch==1.33 ; python_version >= "3.11" and python_version < "4.0"
jsonpointer==3.0.0 ; python_version >= "3.11" and python_version < "4.0"
langchain-community==0.2.7 ; python_version >= "3.11" and python_version < "4.0"
langchain-core==0.2.16 ; python_version >= "3.11" and python_version < "4.0"
langchain-openai==0.1.15 ; python_version >= "3.11" and python_version < "4.0"
langchain-text-splitters==0.2.2 ; python_version >= "3.11" and python_version < "4.0"
langchain==0.2.7 ; python_version >= "3.11" and python_version < "4.0"
langsmith==0.1.85 ; python_version >= "3.11" and python_version < "4.0"
markdown-it-py==3.0.0 ; python_version >= "3.11" and python_version < "4.0"
markupsafe==2.1.5 ; python_version >= "3.11" and python_version < "4.0"
marshmallow==3.21.3 ; python_version >= "3.11" and python_version < "4.0"
mdurl==0.1.2 ; python_version >= "3.11" and python_version < "4.0"
multidict==6.0.5 ; python_version >= "3.11" and python_version < "4.0"
mypy-extensions==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
numpy==1.26.4 ; python_version >= "3.11" and python_version < "4.0"
openai==1.35.13 ; python_version >= "3.11" and python_version < "4.0"
orjson==3.10.5 ; python_version >= "3.11" and python_version < "4.0"
packaging==24.1 ; python_version >= "3.11" and python_version < "4.0"
passlib==1.7.4 ; python_version >= "3.11" and python_version < "4.0"
pathspec==0.12.1 ; python_version >= "3.11" and python_version < "4.0"
pinecone-client==4.1.2 ; python_version >= "3.11" and python_version < "4.0"
pinecone-plugin-interface==0.0.7 ; python_version >= "3.11" and python_version < "4.0"
platformdirs==4.2.2 ; python_version >= "3.11" and python_version < "4.0"
pluggy==1.5.0 ; python_version >= "3.11" and python_version < "4.0"
pyasn1==0.6.0 ; python_version >= "3.11" and python_version < "4.0"
pydantic-core==2.18.4 ; python_version >= "3.11" and python_version < "4.0"
pydantic==2.7.4 ; python_version >= "3.11" and python_version < "4.0"
pygments==2.18.0 ; python_version >= "3.11" and python_version < "4.0"
pyjwt==2.8.0 ; python_version >= "3.11" and python_version < "4.0"
pymupdf==1.24.7 ; python_version >= "3.11" and python_version < "4.0"
pymupdfb==1.24.6 ; python_version >= "3.11" and python_version < "4.0"
pytest==8.3.2 ; python_version >= "3.11" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.11" and python_version < "4.0"
python-jose==3.3.0 ; python_version >= "3.11" and python_version < "4.0"
python-multipart==0.0.9 ; python_version >= "3.11" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.11" and python_version < "4.0"
regex==2024.5.15 ; python_version >= "3.11" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.11" and python_version < "4.0"
rich==13.7.1 ; python_version >= "3.11" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.11" and python_version < "4"
s3transfer==0.10.2 ; python_version >= "3.11" and python_version < "4.0"
shellingham==1.5.4 ; python_version >= "3.11" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
sqlalchemy==2.0.31 ; python_version >= "3.11" and python_version < "4.0"
starlette==0.37.2 ; python_version >= "3.11" and python_version < "4.0"
tenacity==8.5.0 ; python_version >= "3.11" and python_version < "4.0"
tiktoken==0.7.0 ; python_version >= "3.11" and python_version < "4.0"
tqdm==4.66.4 ; python_version >= "3.11" and python_version < "4.0"
typer==0.12.3 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.11" and python_version < "4.0"
typing-inspect==0.9.0 ; python_version >= "3.11" and python_version < "4.0"
ujson==5.10.0 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.2.2 ; python_version >= "3.11" and python_version < "4.0"
uvicorn[standard]==0.30.1 ; python_version >= "3.11" and python_version < "4.0"
uvloop==0.19.0 ; (sys_platform != "win32" and sys_platform != "cygwin") and platform_python_implementation != "PyPy" and python_version >= "3.11" and python_version < "4.0"
watchfiles==0.22.0 ; python_version >= "3.11" and python_version < "4.0"
websockets==12.0 ; python_version >= "3.11" and python_version < "4.0"
yarl==1.9.4 ; python_version >= "3.11" and python_version < "4.0"
